feat: use lazy_static for DbPool
Instead of using a reference to the DbPool to gather connections,
the DbPool is initialized on runtime and defined as a static
using "lazy_static" crate.

This way we are able to initialize the DbPool when values are
read from the .env file and use the same instance for the
application lifecycle.