fix(users): avoid dangling User creation
Currently a user registry is performed in 3 disengaged steps, being:

1. Create a new User instance and store it into the database
2. Create a new Profile instance and store it into the database
3. Create a new Secret instance and store it into the database

All three steps are disengaged and not related to each other, even
when each of these must be achieved in order to successfully register
a new User,otherwise required relations for a User are not creating
thus resulting in an invalid User or dangling User.

In order to fix this behavior many changes were performed, relevant
changes are:

1. Move Auth and Secret definitions from the application layer to the
domain layer as these are strongly engaged to the User entity.

2. Define a BaseRepository trait which must be implemented on Repository-like
traits (supertrait), which relies on SQL transactions.

3. Implement a `register` method which is responsable of beginning a
SQL transaction, consumes UserService, ProfileService and SecretService
with a new family of methods (Defined in step 4) which are transaction
based.

4. Implement `create_tx` methods whihc relies on a SQL transaction
reference to perform queries, thus providing the capability of
either commit or rollback operations as needed.