fix(parcel): parcel filtering based on chat_id
Instead of filter Parcels right on the `write_into_ws_message`
associated function.

The filtering is being done one step before to avoid consuming
resources when a message should not arrive to some user.

Instead of parsing N messages where N is the number of users
connected, we parse U where U is receivers of the message which
could be as much as total users connected or a group of users
(2 or more).